send images to slideshow

+++ src/main/java/com/workspaceit/pmc/api/EventImageApiController.java
@@ -126,6 +126,10 @@ public class EventImageApiController {
         Object principle = authentication.getPrincipal();
         Photographer photographer = (PhotographerUserDetails) principle;
         try {
+            boolean ownership = eventImageService.checkOwnerShipOfImages(imageIds,photographer);
+            if(!ownership){
+                return ResponseEntity.status(HttpStatus.UNPROCESSABLE_ENTITY).body("You don't have access to delete some of the images");
+            }
             boolean result = eventImageService.deleteEventImages(imageIds);
             if(!result){
                 return ResponseEntity.status(HttpStatus.UNPROCESSABLE_ENTITY).body("Something went wrong");
@@ -134,6 +138,15 @@ public class EventImageApiController {
         }catch (EntityNotFound entityNotFound){
             return ResponseEntity.status(HttpStatus.UNPROCESSABLE_ENTITY).body("Entity Not found");
         }
+    }
+
+    @PostMapping("/send-to-slideshow")
+    public ResponseEntity<?> sendImagesToSlideShow(@RequestParam("imageIds") int[] imageIds){
+        boolean result = eventImageService.sendImagesToSlideShow(imageIds);
+        if(!result){
+            return ResponseEntity.status(HttpStatus.UNPROCESSABLE_ENTITY).body("Something went wrong");
+        }
+        return ResponseEntity.status(HttpStatus.OK).body(result);
     }
 
 }
